+++ main.cpp
@@ -1,224 +1,4 @@
-// Include headers
-#include <iostream>
-#include <vector>
-#include <cstdint>
-#include <cstdlib>
-#include <fstream>
-#include <sstream>
-#include <string>
-#include <cstdarg>
-
-// Using directives
-using std::cout;
-using std::cin;
-using std::cerr;
-using std::ifstream;
-using std::endl;
-using std::string;
-using std::vector;
-using std::ostringstream;
-using std::istringstream;
-
-// Global Variables
-const int DIALOGUE_SIZE = 256;
-ostringstream dialogue;
-bool quitGame = false;
-string mapDirectory = "./map/";
-
-// Function Prototypes
-void pushd(const std::string& message, const char* speaker = nullptr);
-void clear();
-
-// Enum Definitions
-enum TileProperties {solid, interactable, has_item};
-enum Tile { 
-  EMPTY = 0, 
-  WALL = 1, 
-  CHAIR = 2, 
-  VERTICAL_DOOR = 3, 
-  HORIZONTAL_DOOR = 4, 
-  TABLE = 5, 
-  LEFT_PANEL = 6, 
-  RIGHT_PANEL = 7
-};
-
-// Classes
-class Map; // Forward declaration
-
-class Player {
-public:
-  int x, y; // Player position
-  Player(int startX, int startY) : x(startX), y(startY) {}
-
-  void playerInput(Map& map); // Now playerInput is a method of the Player class
-};
-
-class Map {
-public:
-  vector<vector<Tile>> mapData; // Map data
-  void drawMap(const Player& player); // Draw the map
-  void getMapData(); // Fetch map data from file
-  int spawn[2]; // Starting position
-};
-
-void Map::drawMap(const Player& player) {
-  clear(); // Clear the console before drawing the map
-  for (int row = 0; row < mapData.size(); ++row) {
-    for (int col = 0; col < mapData[row].size(); ++col) {
-      if (row == player.y && col == player.x) {
-        cout << "Ｏ"; // Draw player
-      } else {
-        switch (mapData[row][col]) {
-          case EMPTY: cout << "　"; break; // Space
-          case WALL: cout << "Ｈ"; break; // Wall
-          case CHAIR: cout << "＃"; break; // Chair
-          case VERTICAL_DOOR: cout << "｜"; break; // Vertical Door
-          case HORIZONTAL_DOOR: cout << "－"; break; // Horizontal Door
-          case TABLE: cout << "＝"; break; // Table/Counter/Bar/Desk
-          case LEFT_PANEL: cout << "［"; break; // Left-facing panel
-          case RIGHT_PANEL: cout << "］"; break; // Right-facing panel
-        }
-      }
-    }
-    cout << endl;
-  }
-}
-
-void Map::getMapData() {
-  mapData.clear(); // Clear existing map data to avoid appending to old data
-  spawn[0] = 0; // Set spawn X position
-  spawn[1] = 0;  // Set spawn Y position
-  
-  ifstream mapFile(mapDirectory + "map.cmap");
-  if (!mapFile.is_open()) {
-    cerr << "Failed to open map file:" << mapDirectory << "map.cmap\n"; 
-    return;
-  }
-  
-  string line;
-  string mapDataSection;
-  
-  // Find the [MAP] section
-  while (getline(mapFile, line)) {
-    if (line == "[MAP]") {
-      // Read the entire [MAP] section as a single string
-      while (getline(mapFile, line)) {
-        if (line.empty() || line[0] == '[') break; // Stop at the next section or empty line
-        mapDataSection += line + " ";
-      }
-
-      // Split the mapSection string into rows based on spaces
-      istringstream mapStream(mapDataSection);
-      string row;
-      while (mapStream >> row) {
-        vector<Tile> rowData;
-        for (char c : row) {
-          if (c == 'a') {
-            // Set spawn position to position of 'a'
-            spawn[0] = rowData.size(); // X position
-            spawn[1] = mapData.size(); // Y position
-          }
-
-          if (c >= '0' && c <= '9') {
-            // Convert numeric characters to enum values
-            rowData.push_back(static_cast<Tile>(c - '0'));
-          }
-        }
-      mapData.push_back(rowData);
-      }
-    }
-
-    if (line == "[SPECIAL]") {
-
-    }
-
-    if (line == "[LANG]") {
-
-    }
-
-    if (line == "[EOF]") {
-      break; // Stop reading when we reach the end
-    }
-
-    if (line.empty() || (line[0] == '/' && line[1] == '/')) {
-      continue; // Skip empty lines or lines that start withh "//"
-    }
-  }
-  mapFile.close(); // Close the file
-}
-
-void Player::playerInput(Map& map) {
-  char cmd;
-  cin >> cmd;
-  
-  Tile target;
-  
-  switch (cmd) {
-    case 'w': 
-    if (y > 0) { // Check if the player isn't at the top edge
-      target = map.mapData[y - 1][x]; // Tile above
-      if (target != WALL) { // If tile above is not a wall, move
-        y--;
-      } else {
-        pushd("You can't go through walls.");
-      }
-    } else {
-      pushd("Going out of bounds isn't allowed.");
-    }
-    break;
-    
-    case 'a':
-    if (x > 0) { // Check if the player isn't at the left edge
-      target = map.mapData[y][x - 1]; // Tile to the left
-      if (target != WALL) { // If tile to left is not a wall, move
-        x--;
-      } else {
-        pushd("You can't go through walls.");
-      }
-    } else {
-      pushd("Going out of bounds isn't allowed.");
-    }
-    break;
-    
-    case 's':
-    if (y < map.mapData.size() - 1) { // Check if the player isn't at the bottom edge
-      target = map.mapData[y + 1][x]; // Tile below
-      if (target != WALL) { // If tile below is not a wall, move
-        y++;
-      } else {
-        pushd("You can't go through walls.");
-      }
-    } else {
-      pushd("Going out of bounds isn't allowed.");
-    }
-    break;
-    
-    case 'd':
-    if (x < map.mapData[y].size() - 1) { // Check if the player isn't at the right edge
-      target = map.mapData[y][x + 1]; // Tile to the right
-      if (target != WALL) { // If tile to right is not a wall, move
-        x++;
-      } else {
-        pushd("You can't go through walls.");
-      }
-    } else {
-      pushd("Going out of bounds isn't allowed.");
-    }
-    break;
-    
-    case 'e':
-    pushd("Interaction hasn't been implemented yet.");
-    break;
-    
-    case 'q':
-    quitGame = true; // Set the quit flag;
-    break;
-    
-    default:
-    pushd("Invalid command.");
-    break;
-  }
-}
+#include "main.hpp"
 
 void clear() { // Clear the console
   cout << "\033[2J\033[1;1H";
